+++ Back/InventorySystem/src/main/java/InventorySystem/SERVICE/MercanciaService.java
@@ -6,7 +6,7 @@ import java.util.List;
 public interface MercanciaService {
     //CRUD - Create, Read, Update, Delete
     public List<MercanciaDTO> listarMercancias();
-    public List<MercanciaDTO> listarPorUsuario(Long idUsuario);
+    public List<MercanciaDTO> listarPorUsuario(int idUsuario);
     public MercanciaDTO guardar(MercanciaDTO mercancia);
     public void eliminar(MercanciaDTO mercancia);
 }
