+++ Back/InventorySystem/src/main/java/InventorySystem/SERVICE/RegistroService.java
@@ -6,7 +6,7 @@ import java.util.List;
 public interface RegistroService {
     //CRUD - Create, Read, Update, Delete
     public List<RegistroDTO> listarRegistros();
-    public List<RegistroDTO> listarPorUsuario(Long idUsuario);
+    public List<RegistroDTO> listarPorUsuario(int idUsuario);
     public List<RegistroDTO> listarPorMercancia(String idMercancia);
     public void guardar(RegistroDTO registro);
 }
