+++ Back/InventorySystem/src/main/java/InventorySystem/SERVICE/UsuarioService.java
@@ -6,7 +6,7 @@ import java.util.List;
 public interface UsuarioService {
     //CRUD - Create, Read, Update, Delete
     public List<UsuarioDTO> listarUsuarios();
-    public List<UsuarioDTO> listarPorCargo(Long idCargo);
+    public List<UsuarioDTO> listarPorCargo(int idCargo);
 //    public void guardar(UsuarioDTO usuario);
 //    public void eliminar(UsuarioDTO usuario);
 //    public UsuarioDTO encontrarUsuario(UsuarioDTO usuario);
